+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @globalfishingwatchapp/fishing-map
 
+## 0.28.6
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/dataviews-client@5.0.1
+  - @globalfishingwatch/layer-composer@4.18.1
+  - @globalfishingwatch/react-hooks@5.1.1
+  - @globalfishingwatch/ui-components@2.16.1
+
+## 0.28.5
+
+### Patch Changes
+
+- 6badcfe5: Update use-legends hook to match fishing-map latest changes
+- f164fc6d: Support mouseEnter and mouseLeave in IconButton
+- Updated dependencies [6badcfe5]
+- Updated dependencies [f164fc6d]
+- Updated dependencies [df1cb0cc]
+  - @globalfishingwatch/react-hooks@5.1.0
+  - @globalfishingwatch/ui-components@2.16.0
+  - @globalfishingwatch/api-types@3.1.0
+
+## 0.28.4
+
+### Patch Changes
+
+- Updated dependencies [d2501b54]
+  - @globalfishingwatch/api-types@3.0.0
+  - @globalfishingwatch/dataviews-client@5.0.0
+  - @globalfishingwatch/react-hooks@5.0.0
+  - @globalfishingwatch/api-client@1.6.8
+
 ## 0.28.3
 
 ### Patch Changes