+++ packages/api-types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/api-types
 
+## 3.1.0
+
+### Minor Changes
+
+- df1cb0cc: support category typing in workspace definition
+
+## 3.0.0
+
+### Major Changes
+
+- d2501b54: move fishing-map logic into dataview-client
+
 ## 2.17.0
 
 ### Minor Changes